codeprompt::files
Function traverse_directory

pub fn traverse_directory(
    root: &Path,
    include: &[String],
    exclude: &[String],
    exclude_priority: bool,
    no_line_numbers: bool,
    relative_paths: bool,
    exclude_from_tree: bool,
    no_codeblock: bool,
    gitignore: bool,
) -> Result<(String, Vec<Value>)>

Starts at the directory root path and traverses the files to build a tree representation.

Arguments

root: The path to the root directory.

include: The include patterns.

exclude: The exclude patterns.

exclude_priority: Whether to give priority to the exclude patterns.

no_line_numbers: Whether to skip adding line numbers to the code sections.

relative_paths: Whether to use relative paths in the file tree.

exclude_from_tree: Whether to exclude files picked up by the exclude patterns from the tree.

no_codeblock: Whether to wrap the code in markdown code blocks.

gitignore: Whether or not to respect the gitignore file.

verbose: Whether to print the glob pattern matching for investigation.

Returns

Result<(String, Vec<serde_json::Value>)>: The string representation of the tree and the JSON representation.